Add skjemavariant field to response DTO
And move request and reponse zod schemas to own file
